pyproject.toml
[project]
name = "pdf-anonymizer"
version = "0.1.0"
description = "A tool to anonymize PDF files using LLMs."
authors = [{ name = "Leonid", email = "leo.gan.57@gmail.com" }]
requires-python = ">=3.9"
dependencies = [
    "python-dotenv",
    "pymupdf4llm",
    "google-genai",
    "typer",
    "rich",
    "ollama",
]

[project.optional-dependencies]
test = [
    "pytest",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
pdf-anonymizer = "pdf_anonymizer.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
